Moved 5 funcs to `scale.py`: ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_value, ui_scale_blit
