+++ scripts/screens/ChangeGenderScreen.py
@@ -19,11 +19,8 @@ from scripts.game_structure.ui_elements import (
 from scripts.utility import (
     get_text_box_theme,
     shorten_text_to_fit,
-    ui_scale_dimensions,
-    ui_scale_value,
-    ui_scale_offset,
 )
-from scripts.utility import ui_scale
+from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_value
 from .Screens import Screens
 from .enums import GameScreen
 from ..game_structure import localization as pronouns
