+++ scripts/screens/ChooseMateScreen.py
@@ -14,11 +14,9 @@ from scripts.game_structure.ui_elements import (
 from scripts.utility import (
     get_personality_compatibility,
     get_text_box_theme,
-    ui_scale,
-    ui_scale_dimensions,
-    ui_scale_offset,
     shorten_text_to_fit,
 )
+from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
 from .enums import GameScreen
 from ..cat.enums import CatCompatibility
