+++ scripts/screens/ClanSettingsScreen.py
@@ -20,10 +20,8 @@ from scripts.game_structure.ui_elements import (
 from scripts.housekeeping.datadir import open_data_dir
 from scripts.utility import (
     get_text_box_theme,
-    ui_scale,
-    ui_scale_dimensions,
-    ui_scale_offset,
 )  # pylint: disable=redefined-builtin
+from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
 from .enums import GameScreen
 from .screens_core.screens_core import rebuild_den_dropdown
