+++ scripts/screens/ClearingScreen.py
@@ -17,9 +17,9 @@ from scripts.game_structure.ui_elements import (
 )
 from scripts.utility import (
     get_text_box_theme,
-    ui_scale,
     shorten_text_to_fit,
 )
+from ..ui.scale import ui_scale
 from .Screens import Screens
 from ..clan_package.settings import get_clan_setting, switch_clan_setting
 from scripts.events_module.short.condition_events import Condition_Events
