+++ scripts/screens/EventsScreen.py
@@ -31,13 +31,15 @@ from scripts.ui.generate_box import BoxStyles, get_box
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
 from scripts.utility import (
-    ui_scale,
     clan_symbol_sprite,
     get_text_box_theme,
     shorten_text_to_fit,
+)
+from scripts.ui.scale import (
+    ui_scale,
     ui_scale_dimensions,
-    ui_scale_value,
     ui_scale_offset,
+    ui_scale_value,
 )
 from scripts.clan_package.get_clan_cats import get_living_clan_cat_count
