+++ scripts/screens/LeaderDenScreen.py
@@ -25,14 +25,13 @@ from scripts.ui.generate_box import get_box, BoxStyles
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
 from scripts.utility import (
-    ui_scale,
     get_text_box_theme,
     get_other_clan_relation,
     get_other_clan,
     clan_symbol_sprite,
     shorten_text_to_fit,
-    ui_scale_dimensions,
 )
+from scripts.ui.scale import ui_scale, ui_scale_dimensions
 from scripts.clan_package.get_clan_cats import (
     find_alive_cats_with_rank,
     get_living_clan_cat_count,
