+++ scripts/ui/windows/changelog.py
@@ -5,7 +5,7 @@ from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import UITextBoxTweaked
 from scripts.housekeeping.version import get_version_info
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale
+from scripts.ui.scale import ui_scale
 from re import search as re_search
