+++ scripts/ui/windows/confirm_display_change.py
@@ -7,7 +7,7 @@ from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import UISurfaceImageButton
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale, ui_scale_offset
+from scripts.ui.scale import ui_scale, ui_scale_offset
 
 if TYPE_CHECKING:
     from scripts.screens.Screens import Screens
