+++ scripts/ui/windows/delete_check.py
@@ -12,7 +12,7 @@ from scripts.housekeeping.datadir import get_save_dir
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale
+from scripts.ui.scale import ui_scale
 
 
 class DeleteCheck(GameWindow):
