+++ scripts/ui/windows/save_as_image.py
@@ -10,7 +10,7 @@ from scripts.game_structure.ui_elements import (
 from scripts.housekeeping.datadir import get_saved_images_dir, open_data_dir
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale
+from scripts.ui.scale import ui_scale
 
 
 class SaveAsImage(GameWindow):
