+++ scripts/ui/windows/save_check.py
@@ -19,7 +19,8 @@ from scripts.game_structure.ui_elements import (
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale, ui_scale_dimensions, quit_game
+from scripts.utility import quit_game
+from scripts.ui.scale import ui_scale, ui_scale_dimensions
 
 
 class SaveCheck(GameWindow):
