+++ scripts/ui/windows/symbol_filter.py
@@ -13,7 +13,7 @@ from scripts.game_structure.ui_elements import (
     UIModifiedScrollingContainer,
 )
 from scripts.ui.windows.base_window import GameWindow
-from scripts.utility import ui_scale
+from scripts.ui.scale import ui_scale
 
 
 class SymbolFilterWindow(GameWindow):
