+++ scripts/utility.py
@@ -14,7 +14,7 @@ from itertools import combinations
 from math import floor
 from random import choice, choices, randint, random, sample, randrange, getrandbits
 from sys import exit as sys_exit
-from typing import List, Tuple, TYPE_CHECKING, Type, Union, Optional
+from typing import List, TYPE_CHECKING, Type, Union, Optional
 
 import i18n
 import pygame
@@ -31,6 +31,7 @@ from scripts.game_structure.localization import (
     determine_plural_pronouns,
     get_lang_config,
 )
+from scripts.ui.scale import ui_scale_dimensions
 
 logger = logging.getLogger(__name__)
 from scripts.game_structure import image_cache, localization, constants
@@ -2332,89 +2333,6 @@ def shorten_text_to_fit(
 # ---------------------------------------------------------------------------- #
 #                                    Sprites                                   #
 # ---------------------------------------------------------------------------- #
-
-
-def ui_scale(rect: pygame.Rect):
-    """
-    Scales a pygame.Rect appropriately for the UI scaling currently in use.
-    :param rect: a pygame.Rect
-    :return: the same pygame.Rect, scaled for the current UI.
-    """
-    # offset can be negative to allow for correct anchoring
-    rect[0] = floor(rect[0] * scripts.game_structure.screen_settings.screen_scale)
-    rect[1] = floor(rect[1] * scripts.game_structure.screen_settings.screen_scale)
-    # if the dimensions are negative, it's dynamically scaled, ignore
-    rect[2] = (
-        floor(rect[2] * scripts.game_structure.screen_settings.screen_scale)
-        if rect[2] > 0
-        else rect[2]
-    )
-    rect[3] = (
-        floor(rect[3] * scripts.game_structure.screen_settings.screen_scale)
-        if rect[3] > 0
-        else rect[3]
-    )
-
-    return rect
-
-
-def ui_scale_dimensions(dim: Tuple[int, int]):
-    """
-    Use to scale the dimensions of an item - WILL IGNORE NEGATIVE VALUES
-    :param dim: The dimensions to scale
-    :return: The scaled dimensions
-    """
-    return (
-        (
-            floor(dim[0] * scripts.game_structure.screen_settings.screen_scale)
-            if dim[0] > 0
-            else dim[0]
-        ),
-        (
-            floor(dim[1] * scripts.game_structure.screen_settings.screen_scale)
-            if dim[1] > 0
-            else dim[1]
-        ),
-    )
-
-
-def ui_scale_offset(coords: Tuple[int, int]):
-    """
-    Use to scale the offset of an item (i.e. the first 2 values of a pygame.Rect).
-    Not to be confused with ui_scale_blit.
-    :param coords: The coordinates to scale
-    :return: The scaled coordinates
-    """
-    return (
-        floor(coords[0] * scripts.game_structure.screen_settings.screen_scale),
-        floor(coords[1] * scripts.game_structure.screen_settings.screen_scale),
-    )
-
-
-def ui_scale_value(val: int):
-    """
-    Use to scale a single value according to the UI scale. If you need this one,
-    you're probably doing something unusual. Try to avoid where possible.
-    :param val: The value to scale
-    :return: The scaled value
-    """
-    return floor(val * scripts.game_structure.screen_settings.screen_scale)
-
-
-def ui_scale_blit(coords: Tuple[int, int]):
-    """
-    Use to scale WHERE to blit an item, not the SIZE of it. (0, 0) is the top left corner of the pygame_gui managed window,
-    this adds the offset from fullscreen etc. to make it blit in the right place. Not to be confused with ui_scale_offset.
-    :param coords: The coordinates to blit to
-    :return: The scaled, correctly offset coordinates to blit to.
-    """
-    return floor(
-        coords[0] * scripts.game_structure.screen_settings.screen_scale
-        + scripts.game_structure.screen_settings.offset[0]
-    ), floor(
-        coords[1] * scripts.game_structure.screen_settings.screen_scale
-        + scripts.game_structure.screen_settings.offset[1]
-    )
 
 
 def update_sprite(cat):
